move track and queue handling to sep file
